fix: add missing docblock and defensive int cast in Ookla actions
Add the standard "Handle the action." docblock to RunSpeedtest::handle()
to match this codebase's PHPDoc convention. Cast config('speedtest.retry_times')
to int in RetrySpeedtest so a non-numeric or string config value can't
silently break the retry-times comparison.

+++ app/Actions/Ookla/RetrySpeedtest.php
@@ -14,7 +14,7 @@ class RetrySpeedtest
      */
     public function handle(Result $result, int $attempt): void
     {
-        $retryTimes = config('speedtest.retry_times');
+        $retryTimes = (int) config('speedtest.retry_times');
 
         // Retrying is disabled entirely.
         if ($retryTimes === 0) {
